move Cursor from tests under velox/exec
